refactor: Check if fzf is already installed

+++ fzf/install.sh
@@ -10,18 +10,21 @@ source "$DOTFILES_ROOT/lib/logger.sh"
 
 info 'Installing fzf'
 
-if ! command -v brew > /dev/null; then
-  user 'Homebrew not found. Would you like to install Homebrew first? (y/n)'
-  read -r -p '> ' install_brew
+if ! command -v fzf > /dev/null; then
+  if ! command -v brew > /dev/null; then
+    user 'Homebrew not found. Would you like to install Homebrew first? (y/n)'
+    read -r -p '> ' install_brew
 
-  if [[ "$install_brew" =~ ^[Yy]$ ]]; then
-    source "$DOTFILES_ROOT/homebrew/install.sh"
-    source "$DOTFILES_ROOT/homebrew/path.zsh"
-  else
-    fail 'Homebrew is required to install fzf.'
+    if [[ "$install_brew" =~ ^[Yy]$ ]]; then
+      source "$DOTFILES_ROOT/homebrew/install.sh"
+      source "$DOTFILES_ROOT/homebrew/path.zsh"
+    else
+      fail 'Homebrew is required to install fzf.'
+    fi
   fi
+  brew install fzf
+
+  success 'fzf installed'
+else
+  success 'fzf is already installed'
 fi
-
-brew install fzf
-
-success 'fzf installed'
